refactor: Move Index tests to base_index
To make it easier to turn git2 into an optional dep.
